+++ src/main/java/com/verdantartifice/primalmagick/client/gui/radial/DrawingContext.java
@@ -1,0 +1,61 @@
+package com.verdantartifice.primalmagick.client.gui.radial;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.renderer.entity.ItemRenderer;
+
+public class DrawingContext {
+    public final int width;
+    public final int height;
+    public final float x;
+    public final float y;
+    public final float z;
+    public final Font fontRenderer;
+    public final ItemRenderer itemRenderer;
+    public final PoseStack matrixStack;
+    public final IDrawingHelper drawingHelper;
+
+    public DrawingContext(PoseStack matrixStack, int width, int height, float x, float y, float z, Font fontRenderer, ItemRenderer itemRenderer, IDrawingHelper drawingHelper)
+    {
+        this.matrixStack = matrixStack;
+        this.width = width;
+        this.height = height;
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.fontRenderer = fontRenderer;
+        this.itemRenderer = itemRenderer;
+        this.drawingHelper = drawingHelper;
+    }
+}
+
+/*
+ Note: This code was adapted from David Quintana's implementation in Tool Belt.
+ Below is the copyright notice.
+ 
+ Copyright (c) 2015, David Quintana <gigaherz@gmail.com>
+ All rights reserved.
+
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions are met:
+     * Redistributions of source code must retain the above copyright
+       notice, this list of conditions and the following disclaimer.
+     * Redistributions in binary form must reproduce the above copyright
+       notice, this list of conditions and the following disclaimer in the
+       documentation and/or other materials provided with the distribution.
+     * Neither the name of the author nor the
+       names of the contributors may be used to endorse or promote products
+       derived from this software without specific prior written permission.
+
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ DISCLAIMED. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
+ DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */